Added Sign In page and New account page
Landing page is intended to be the sign-in page.  If the user does not
have an account they will not be allowed to log in and will have to make
a new account.  When a new account is made they will be redirected to
the sign in page.  A successful login will take them to the MainPage