Username & Password reset finished
The user can now reset his username and password using the respective
buttons. The changes are made both to the sessionstorage version of the
user, and the permanent version in localstorage. If the user has posted
any books, the "seller" of those books will change to reflect the user's
changed contact information.